requirements.txt
functions-framework==3.*
requests==2.26.0
python-dotenv==1.0.0